requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pytest
httpx
google-genai
